[Fix] 서버 재시작 시, job 다시 define
