+++ src/loaders/db.ts
@@ -2,6 +2,8 @@ import mongoose from "mongoose";
 import config from "../config";
 import * as admin from "firebase-admin";
 import serviceAccount from "../config/firebase-admin.json";
+import User from "../models/User";
+import agenda from "./agenda";
 
 const connectDB = async () => {
   let firebase;
@@ -10,7 +12,6 @@ const connectDB = async () => {
       firebase = admin.initializeApp({
         credential: admin.credential.cert(serviceAccount as admin.ServiceAccount),
       });
-      // console.log("SDK 초기화 완료!");
     } else {
       firebase = admin.app();
       console.log("FCM 준비 완료!");
@@ -20,7 +21,51 @@ const connectDB = async () => {
 
     mongoose.set("autoCreate", true);
 
-    // console.log("Mongoose Connected ...");
+    // 푸시알림
+    const user = await User.find({ isActive: true });
+
+    for (let i = 0; i < user.length; i++) {
+      const alarms = {
+        android: {
+          data: {
+            title: "test recordream",
+            body: "테스트입니다.",
+          },
+        },
+        apns: {
+          payload: {
+            aps: {
+              contentAvailable: true,
+              alert: {
+                title: "test recordream",
+                body: "테스트입니다.",
+              },
+            },
+          },
+        },
+        tokens: user[i].fcmTokens,
+      };
+
+      agenda.define("push" + `${user[i]._id}`, async (job: any, done: any) => {
+        admin
+          .messaging()
+          .sendMulticast(alarms)
+          .then(function (res) {
+            console.log("Sent message result: ", res);
+          });
+        job.repeatEvery("24 hours").save();
+        done();
+      });
+      agenda.start();
+    }
+
+    const graceful = () => {
+      agenda.stop();
+      process.exit(0);
+    };
+
+    process.on("SIGTERM", graceful);
+    process.on("SIGINT", graceful);
   } catch (err) {
     console.error(err);
     process.exit(1);
